+++ src/snippets/dto/listAllEntities.dto.ts
