+++ src/snippets/exceptions/not-found.exception.ts
