+++ src/snippets/interfaces/snippets.interface.ts
