+++ src/snippets/pipes/user-by-id.pipe.ts
