+++ src/snippets/validation/snippetValidation.js
