+++ src/snippets/validation/validateAuth.ts
