--- src/src/app.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class AppService {
-  // eslint-disable-next-line class-methods-use-this
-  getHello(): string {
-    return 'Hello World!';
-  }
-}
