--- src/src/main.ts
@@ -1,20 +1,0 @@
-/* eslint-disable import/no-import-module-exports */
-import { NestFactory } from '@nestjs/core';
-import { NestExpressApplication } from '@nestjs/platform-express';
-import { join } from 'path';
-import { AppModule } from './app.module';
-
-declare const module: any;
-
-async function bootstrap() {
-  const app = await NestFactory.create<NestExpressApplication>(AppModule);
-  app.useStaticAssets(join(__dirname, '..', 'assets'));
-  app.setBaseViewsDir(join(__dirname, '..', 'views'));
-  app.setViewEngine('pug');
-  await app.listen(process.env.PORT || 3000);
-  if (module.hot) {
-    module.hot.accept();
-    module.hot.dispose(() => app.close());
-  }
-}
-bootstrap();
